+++ src/components/Checkbox/Checkbox.tsx
@@ -6,8 +6,8 @@ import Check from "../Icon/Icons/components/Check";
 import Remove from "../Icon/Icons/components/Remove";
 import { backwardCompatibilityForProperties } from "../../helpers/backwardCompatibilityForProperties";
 import { useSupportFirefoxLabelClick } from "./hooks/useSupportFirefoxLabelClick";
-import useMergeRefs from "../../hooks/useMergeRefs";
-import { VibeComponentProps } from "../../types";
+import useMergeRef from "../../hooks/useMergeRef";
+import { VibeComponent, VibeComponentProps } from "../../types";
 import { getTestId } from "../../tests/test-ids-utils";
 import { ComponentDefaultTestId } from "../../tests/constants";
 import Text from "../Text/Text";
@@ -50,7 +50,7 @@ export interface CheckBoxProps extends VibeComponentProps {
   tabIndex?: number;
 }
 
-const Checkbox: React.FC<CheckBoxProps> = forwardRef(
+const Checkbox: VibeComponent<CheckBoxProps, HTMLInputElement> = forwardRef(
   (
     {
       className,
@@ -74,10 +74,11 @@ const Checkbox: React.FC<CheckBoxProps> = forwardRef(
     },
     ref
   ) => {
-    const iconContainerRef = useRef<HTMLDivElement>(null);
     const inputRef = useRef<HTMLInputElement>(null);
-    const mergedInputRef = useMergeRefs({ refs: [ref, inputRef] });
+    const mergedInputRef = useMergeRef(ref, inputRef);
+    const iconContainerRef = useRef<HTMLDivElement>(null);
     const overrideClassName = backwardCompatibilityForProperties([className, componentClassName]);
+
     const onMouseUpCallback = useCallback(() => {
       const input = inputRef.current;
       if (!input) return;