+++ src/components/Menu/MenuItem/MenuItem.tsx
@@ -1,14 +1,14 @@
 /* eslint-disable react/jsx-props-no-spreading */
-import cx from "classnames";
 import React, { ForwardedRef, forwardRef, useCallback, useEffect, useLayoutEffect, useMemo, useRef } from "react";
+import cx from "classnames";
+import { isFunction } from "lodash-es";
 import { ComponentDefaultTestId, getTestId } from "../../../tests/test-ids-utils";
 import { DialogPosition } from "../../../constants/positions";
-import { isFunction } from "lodash-es";
 import Text from "../../Text/Text";
 import Tooltip from "../../../components/Tooltip/Tooltip";
 import Icon from "../../../components/Icon/Icon";
 import DialogContentContainer from "../../../components/DialogContentContainer/DialogContentContainer";
-import useMergeRefs from "../../../hooks/useMergeRefs";
+import useMergeRef from "../../../hooks/useMergeRef";
 import useIsOverflowing from "../../../hooks/useIsOverflowing/useIsOverflowing";
 import usePopover from "../../../hooks/usePopover";
 import { backwardCompatibilityForProperties } from "../../../helpers/backwardCompatibilityForProperties";
@@ -18,8 +18,8 @@ import { SubIcon, VibeComponent, VibeComponentProps, withStaticProps } from "../
 import { IconType } from "../../Icon/IconConstants";
 import { TAB_INDEX_FOCUS_WITH_JS_ONLY, TooltipPosition } from "./MenuItemConstants";
 import { CloseMenuOption, MenuChild } from "../Menu/MenuConstants";
-import styles from "./MenuItem.module.scss";
 import Label from "../../Label/Label";
+import styles from "./MenuItem.module.scss";
 import { DropdownChevronRight } from "../../Icon/Icons";
 import { IconButton } from "../../index";
 import Divider from "../../Divider/Divider";
@@ -141,6 +141,7 @@ const MenuItem: VibeComponent<MenuItemProps> & {
     const popperElement = popperElementRef.current;
     const referenceElement = referenceElementRef.current;
     const childElement = childRef.current;
+    const mergedRef = useMergeRef(ref, referenceElementRef);
 
     const isMouseEnterMenuItem = useIsMouseEnter({ ref: referenceElementRef });
     const isMouseEnterIconButton = useIsMouseEnter({ ref: iconButtonElementRef });