+++ src/hooks/useIsMouseEnter.ts
@@ -1,7 +1,8 @@
 import { useCallback, useState, RefObject } from "react";
 import useEventListener from "./useEventListener";
 
-export default function useMergeRefs({ ref }: { ref: RefObject<HTMLElement> }) {
+// TODO can be replaced with useIsMouseOver
+export default function useIsMouseEnter({ ref }: { ref: RefObject<HTMLElement> }) {
   const [isHovered, setIsHover] = useState<boolean>(false);
 
   const setHovered = useCallback(